+++ host/driver_online/conv_fwd_driver_online.cpp
@@ -12,8 +12,8 @@
 #include "conv_common.hpp"
 #include "host_conv.hpp"
 #include "device_tensor.hpp"
-#include "olc_device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw.hpp"
-#include "olc_device_dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw.hpp"
+#include "online_device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw.hpp"
+#include "online_device_dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw.hpp"
 
 #define USE_CONV_FWD_V4R4_NCHW 1
 #define USE_CONV_FWD_V4R5_NCHW 1
@@ -214,9 +214,9 @@ int main(int argc, char* argv[])
         tunable_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw* tunable =
             &default_tunable_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw;
 
-        device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc<in_data_t,
-                                                                                 acc_data_t,
-                                                                                 out_data_t>(
+        online_device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw<in_data_t,
+                                                                                    acc_data_t,
+                                                                                    out_data_t>(
             handle,
             tmp[I0],
             tmp[I1],
@@ -246,9 +246,9 @@ int main(int argc, char* argv[])
         tunable_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw* tunable =
             &default_tunable_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw;
 
-        device_dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw_olc<in_data_t,
-                                                                                 acc_data_t,
-                                                                                 out_data_t>(
+        online_device_dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw<in_data_t,
+                                                                                    acc_data_t,
+                                                                                    out_data_t>(
             handle,
             tmp[I0],
             tmp[I1],
