+++ host/driver_online/include/online_device_dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw.hpp
@@ -2,14 +2,14 @@
 #include "host_tensor.hpp"
 #include "dynamic_tensor_descriptor.hpp"
 #include "dynamic_tensor_descriptor_helper.hpp"
-#include "transform_forward_convolution_into_gemm_v4r4_nchw_kcyx_nkhw.hpp"
+#include "transform_forward_convolution_into_gemm_v4r5_nchw_kcyx_nkhw.hpp"
 
-#include "olc_driver_common.hpp"
+#include "online_driver_common.hpp"
 #include "conv_tunables.hpp"
 
 #include "handle.hpp"
 
-namespace detail_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw {
+namespace detail_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw {
 
 template <typename TInWei, typename TAcc, typename TOut>
 static std::string get_network_config_string_from_types()
@@ -24,13 +24,13 @@ static std::string get_network_config_string_from_types()
 };
 
 static std::string
-get_network_config_string_from_tunable(const tunable_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw* pt)
+get_network_config_string_from_tunable(const tunable_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw* pt)
 {
     std::string out("TUN_");
 
     out += std::to_string(pt->BlockSize) + "_";
 
-    out += std::to_string(pt->MPerBlock) + "x" + std::to_string(pt->NPerBlock) + "x" +
+    out += std::to_string(pt->GM1PerBlockGM11) + "x" + std::to_string(pt->GN1PerBlockGN11) + "x" +
            std::to_string(pt->KPerBlock) + "_";
     out += std::to_string(pt->M1PerThread) + "x" + std::to_string(pt->N1PerThread) + "x" +
            std::to_string(pt->KPerThread) + "_";
@@ -39,46 +39,54 @@ get_network_config_string_from_tunable(const tunable_dyn_conv_fwd_v4r4_nchw_kcyx
            std::to_string(pt->M1N1ThreadClusterM11) + "x" +
            std::to_string(pt->M1N1ThreadClusterN11) + "_";
 
-    out += std::to_string(pt->ABlockTransferThreadSliceLengths_K_M0_M1[0]) + "x" +
-           std::to_string(pt->ABlockTransferThreadSliceLengths_K_M0_M1[1]) + "x" +
-           std::to_string(pt->ABlockTransferThreadSliceLengths_K_M0_M1[2]) + "_";
+    out += std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[0]) + "x" +
+           std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[1]) + "x" +
+           std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[2]) + "x" +
+           std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[3]) + "_";
 
-    out += std::to_string(pt->ABlockTransferThreadClusterLengths_K_M0_M1[0]) + "x" +
-           std::to_string(pt->ABlockTransferThreadClusterLengths_K_M0_M1[1]) + "x" +
-           std::to_string(pt->ABlockTransferThreadClusterLengths_K_M0_M1[2]) + "_";
+    out += std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[0]) + "x" +
+           std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[1]) + "x" +
+           std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[2]) + "x" +
+           std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[3]) + "_";
 
     out += std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[0]) + "x" +
            std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[1]) + "x" +
-           std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[2]) + "_";
+           std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[2]) + "x" +
+           std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[3]) + "_";
 
     out += std::to_string(pt->ABlockTransferSrcAccessOrder[0]) + "x" +
            std::to_string(pt->ABlockTransferSrcAccessOrder[1]) + "x" +
-           std::to_string(pt->ABlockTransferSrcAccessOrder[2]) + "_";
+           std::to_string(pt->ABlockTransferSrcAccessOrder[2]) + "x" +
+           std::to_string(pt->ABlockTransferSrcAccessOrder[3]) + "_";
 
     out += std::to_string(pt->ABlockTransferSrcVectorDim) + "_";
     out += std::to_string(pt->ABlockTransferSrcScalarPerVector) + "_";
-    out += std::to_string(pt->ABlockTransferDstScalarPerVector_M1) + "_";
+    out += std::to_string(pt->ABlockTransferDstScalarPerVector_GM11) + "_";
     out += std::to_string(pt->AThreadTransferSrcResetCoordinateAfterRun) + "_";
 
-    out += std::to_string(pt->BBlockTransferThreadSliceLengths_K_N0_N1[0]) + "x" +
-           std::to_string(pt->BBlockTransferThreadSliceLengths_K_N0_N1[1]) + "x" +
-           std::to_string(pt->BBlockTransferThreadSliceLengths_K_N0_N1[2]) + "_";
+    out += std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[0]) + "x" +
+           std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[1]) + "x" +
+           std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[2]) + "x" +
+           std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[3]);
 
-    out += std::to_string(pt->BBlockTransferThreadClusterLengths_K_N0_N1[0]) + "x" +
-           std::to_string(pt->BBlockTransferThreadClusterLengths_K_N0_N1[1]) + "x" +
-           std::to_string(pt->BBlockTransferThreadClusterLengths_K_N0_N1[2]) + "_";
+    out += std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[0]) + "x" +
+           std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[1]) + "x" +
+           std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[2]) + "x" +
+           std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[3]) + "_";
 
     out += std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[0]) + "x" +
            std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[1]) + "x" +
-           std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[2]) + "_";
+           std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[2]) + "x" +
+           std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[3]) + "_";
 
     out += std::to_string(pt->BBlockTransferSrcAccessOrder[0]) + "x" +
            std::to_string(pt->BBlockTransferSrcAccessOrder[1]) + "x" +
-           std::to_string(pt->BBlockTransferSrcAccessOrder[2]) + "_";
+           std::to_string(pt->BBlockTransferSrcAccessOrder[2]) + "x" +
+           std::to_string(pt->BBlockTransferSrcAccessOrder[3]) + "_";
 
     out += std::to_string(pt->BBlockTransferSrcVectorDim) + "_";
     out += std::to_string(pt->BBlockTransferSrcScalarPerVector) + "_";
-    out += std::to_string(pt->BBlockTransferDstScalarPerVector_N1) + "_";
+    out += std::to_string(pt->BBlockTransferDstScalarPerVector_GN11) + "_";
     out += std::to_string(pt->BThreadTransferSrcResetCoordinateAfterRun) + "_";
 
     out += std::to_string(pt->CThreadTransferSrcDstAccessOrder[0]) + "x" +
@@ -107,14 +115,14 @@ static std::string get_definition_string_from_types()
 };
 
 static std::string
-get_definition_string_from_tunable(const tunable_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw* pt)
+get_definition_string_from_tunable(const tunable_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw* pt)
 {
     std::string out;
 
     out += " -DCK_PARAM_BlockSize=" + std::to_string(pt->BlockSize);
 
-    out += " -DCK_PARAM_MPerBlock=" + std::to_string(pt->MPerBlock) +
-           " -DCK_PARAM_NPerBlock=" + std::to_string(pt->NPerBlock) +
+    out += " -DCK_PARAM_GM1PerBlockGM11=" + std::to_string(pt->GM1PerBlockGM11) +
+           " -DCK_PARAM_GN1PerBlockGN11=" + std::to_string(pt->GN1PerBlockGN11) +
            " -DCK_PARAM_KPerBlock=" + std::to_string(pt->KPerBlock);
     out += " -DCK_PARAM_M1PerThread=" + std::to_string(pt->M1PerThread) +
            " -DCK_PARAM_N1PerThread=" + std::to_string(pt->N1PerThread) +
@@ -125,61 +133,69 @@ get_definition_string_from_tunable(const tunable_dyn_conv_fwd_v4r4_nchw_kcyx_nkh
            " -DCK_PARAM_M1N1ThreadClusterM11=" + std::to_string(pt->M1N1ThreadClusterM11) +
            " -DCK_PARAM_M1N1ThreadClusterN11=" + std::to_string(pt->M1N1ThreadClusterN11);
 
-    out += " -DCK_PARAM_ABlockTransferThreadSliceLengths_K_M0_M1=" +
-           std::to_string(pt->ABlockTransferThreadSliceLengths_K_M0_M1[0]) + "," +
-           std::to_string(pt->ABlockTransferThreadSliceLengths_K_M0_M1[1]) + "," +
-           std::to_string(pt->ABlockTransferThreadSliceLengths_K_M0_M1[2]);
+    out += " -DCK_PARAM_ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11=" +
+           std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[0]) + "," +
+           std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[1]) + "," +
+           std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[2]) + "," +
+           std::to_string(pt->ABlockTransferThreadSliceLengths_GK_GM0_GM10_GM11[3]);
 
-    out += " -DCK_PARAM_ABlockTransferThreadClusterLengths_K_M0_M1=" +
-           std::to_string(pt->ABlockTransferThreadClusterLengths_K_M0_M1[0]) + "," +
-           std::to_string(pt->ABlockTransferThreadClusterLengths_K_M0_M1[1]) + "," +
-           std::to_string(pt->ABlockTransferThreadClusterLengths_K_M0_M1[2]);
+    out += " -DCK_PARAM_ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11=" +
+           std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[0]) + "," +
+           std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[1]) + "," +
+           std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[2]) + "," +
+           std::to_string(pt->ABlockTransferThreadClusterLengths_GK_GM0_GM10_GM11[3]);
 
     out += " -DCK_PARAM_ABlockTransferThreadClusterArrangeOrder=" +
            std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[0]) + "," +
            std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[1]) + "," +
-           std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[2]);
+           std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[2]) + "," +
+           std::to_string(pt->ABlockTransferThreadClusterArrangeOrder[3]);
 
     out += " -DCK_PARAM_ABlockTransferSrcAccessOrder=" +
            std::to_string(pt->ABlockTransferSrcAccessOrder[0]) + "," +
            std::to_string(pt->ABlockTransferSrcAccessOrder[1]) + "," +
-           std::to_string(pt->ABlockTransferSrcAccessOrder[2]);
+           std::to_string(pt->ABlockTransferSrcAccessOrder[2]) + "," +
+           std::to_string(pt->ABlockTransferSrcAccessOrder[3]);
 
     out +=
         " -DCK_PARAM_ABlockTransferSrcVectorDim=" + std::to_string(pt->ABlockTransferSrcVectorDim);
     out += " -DCK_PARAM_ABlockTransferSrcScalarPerVector=" +
            std::to_string(pt->ABlockTransferSrcScalarPerVector);
-    out += " -DCK_PARAM_ABlockTransferDstScalarPerVector_M1=" +
-           std::to_string(pt->ABlockTransferDstScalarPerVector_M1);
+    out += " -DCK_PARAM_ABlockTransferDstScalarPerVector_GM11=" +
+           std::to_string(pt->ABlockTransferDstScalarPerVector_GM11);
     out += " -DCK_PARAM_AThreadTransferSrcResetCoordinateAfterRun=" +
            std::to_string(pt->AThreadTransferSrcResetCoordinateAfterRun);
 
-    out += " -DCK_PARAM_BBlockTransferThreadSliceLengths_K_N0_N1=" +
-           std::to_string(pt->BBlockTransferThreadSliceLengths_K_N0_N1[0]) + "," +
-           std::to_string(pt->BBlockTransferThreadSliceLengths_K_N0_N1[1]) + "," +
-           std::to_string(pt->BBlockTransferThreadSliceLengths_K_N0_N1[2]);
+    out += " -DCK_PARAM_BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11=" +
+           std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[0]) + "," +
+           std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[1]) + "," +
+           std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[2]) + "," +
+           std::to_string(pt->BBlockTransferThreadSliceLengths_GK_GN0_GN10_GN11[3]);
 
-    out += " -DCK_PARAM_BBlockTransferThreadClusterLengths_K_N0_N1=" +
-           std::to_string(pt->BBlockTransferThreadClusterLengths_K_N0_N1[0]) + "," +
-           std::to_string(pt->BBlockTransferThreadClusterLengths_K_N0_N1[1]) + "," +
-           std::to_string(pt->BBlockTransferThreadClusterLengths_K_N0_N1[2]);
+    out += " -DCK_PARAM_BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11=" +
+           std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[0]) + "," +
+           std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[1]) + "," +
+           std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[2]) + "," +
+           std::to_string(pt->BBlockTransferThreadClusterLengths_GK_GN0_GN10_GN11[3]);
 
     out += " -DCK_PARAM_BBlockTransferThreadClusterArrangeOrder=" +
            std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[0]) + "," +
            std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[1]) + "," +
-           std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[2]);
+           std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[2]) + "," +
+           std::to_string(pt->BBlockTransferThreadClusterArrangeOrder[3]);
 
     out += " -DCK_PARAM_BBlockTransferSrcAccessOrder=" +
            std::to_string(pt->BBlockTransferSrcAccessOrder[0]) + "," +
            std::to_string(pt->BBlockTransferSrcAccessOrder[1]) + "," +
-           std::to_string(pt->BBlockTransferSrcAccessOrder[2]);
+           std::to_string(pt->BBlockTransferSrcAccessOrder[2]) + "," +
+           std::to_string(pt->BBlockTransferSrcAccessOrder[3]);
 
     out +=
         " -DCK_PARAM_BBlockTransferSrcVectorDim=" + std::to_string(pt->BBlockTransferSrcVectorDim);
     out += " -DCK_PARAM_BBlockTransferSrcScalarPerVector=" +
            std::to_string(pt->BBlockTransferSrcScalarPerVector);
-    out += " -DCK_PARAM_BBlockTransferDstScalarPerVector_N1=" +
-           std::to_string(pt->BBlockTransferDstScalarPerVector_N1);
+    out += " -DCK_PARAM_BBlockTransferDstScalarPerVector_GN11=" +
+           std::to_string(pt->BBlockTransferDstScalarPerVector_GN11);
     out += " -DCK_PARAM_BThreadTransferSrcResetCoordinateAfterRun=" +
            std::to_string(pt->BThreadTransferSrcResetCoordinateAfterRun);
 
@@ -199,7 +215,7 @@ get_definition_string_from_tunable(const tunable_dyn_conv_fwd_v4r4_nchw_kcyx_nkh
     return (out);
 };
 
-} // namespace detail_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw
+} // namespace detail_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw
 
 template <typename TInWei,
           typename TAcc,
@@ -211,7 +227,7 @@ template <typename TInWei,
           typename ConvDilations,
           typename InLeftPads,
           typename InRightPads>
-void device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc(
+void online_device_dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw(
     olCompile::Handle* handle,
     const InLengths& in_n_c_hi_wi_lengths,
     const WeiLengths& wei_k_c_y_x_lengths,
@@ -223,14 +239,16 @@ void device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc(
     const Tensor<TInWei>& in_n_c_hi_wi,
     const Tensor<TInWei>& wei_k_c_y_x,
     Tensor<TOut>& out_n_k_ho_wo,
-    const tunable_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw* tunable,
+    const tunable_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw* tunable,
     ck::index_t nrepeat)
 {
     using namespace ck;
-    using namespace detail_dyn_conv_fwd_v4r4_nchw_kcyx_nkhw;
+    using namespace detail_dyn_conv_fwd_v4r5_nchw_kcyx_nkhw;
     using size_t = std::size_t;
 
-    /////////////////////////////////////////////////////////////////////////////////////////////////////////////
+    constexpr index_t N0 = 4; // this could not be a tunable so far
+
+    ////////////////////////////////////////////////////////////////////////////////////////////////////////////
     // The follow codes are only used for computing the grid_size, hasMainKBlockLoop,
     // hasDoubleTailKBlockLoop
 
@@ -246,24 +264,27 @@ void device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc(
     const auto out_n_k_ho_wo_desc =
         make_dynamic_naive_tensor_descriptor_packed_v2(out_n_k_ho_wo_lengths);
 
-    const auto descs =
-        transform_forward_convolution_into_gemm_v4r4_nchw_kcyx_nkhw_pad(wei_k_c_y_x_desc,
-                                                                        in_n_c_hi_wi_desc,
-                                                                        out_n_k_ho_wo_desc,
-                                                                        conv_strides,
-                                                                        conv_dilations,
-                                                                        in_left_pads,
-                                                                        in_right_pads);
-    const auto a_k_m_grid_desc = descs[I0];
-    const auto c_m_n_grid_desc = descs[I2];
-    const auto M               = c_m_n_grid_desc.GetLength(I0);
-    const auto N               = c_m_n_grid_desc.GetLength(I1);
-    const auto K               = a_k_m_grid_desc.GetLength(I0);
+    const auto descs = transform_forward_convolution_into_contraction_v4r5_nchw_kcyx_nkhw_pad<N0>(
+        wei_k_c_y_x_desc,
+        in_n_c_hi_wi_desc,
+        out_n_k_ho_wo_desc,
+        conv_strides,
+        conv_dilations,
+        in_left_pads,
+        in_right_pads);
 
-    const index_t grid_size            = (M / tunable->MPerBlock) * (N / tunable->NPerBlock);
-    const bool hasMainKBlockLoop       = ((K + tunable->KPerBlock) / (2 * tunable->KPerBlock) > 1);
-    const bool hasDoubleTailKBlockLoop = ((K / tunable->KPerBlock) % 2 == 0);
-    /////////////////////////////////////////////////////////////////////////////////////////////////////////////
+    const auto a_gk_gm0_gm1_grid_desc      = descs[I0];
+    const auto c_gm0_gm1_gn0_gn1_grid_desc = descs[I2];
+
+    const auto GM1 = c_gm0_gm1_gn0_gn1_grid_desc.GetLength(I1);
+    const auto GN1 = c_gm0_gm1_gn0_gn1_grid_desc.GetLength(I3);
+    const auto GK  = a_gk_gm0_gm1_grid_desc.GetLength(I0);
+
+    const index_t grid_size = (GM1 / tunable->GM1PerBlockGM11) * (GN1 / tunable->GN1PerBlockGN11);
+    const bool hasMainKBlockLoop       = ((GK + tunable->KPerBlock) / (2 * tunable->KPerBlock) > 1);
+    const bool hasDoubleTailKBlockLoop = ((GK / tunable->KPerBlock) % 2 == 0);
+
+    ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 
     // these buffers are usually provided by the user application
     DeviceMem in_n_c_hi_wi_dev_buf(sizeof(TInWei) * in_n_c_hi_wi.mDesc.GetElementSpace());
@@ -278,32 +299,32 @@ void device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc(
     // workspace API
     DeviceMem workspace_buf(4096);
 
-    void* a_k_m0_m1_grid_desc_dev_buf = workspace_buf.GetDeviceBuffer();
-    void* b_k_n0_n1_grid_desc_dev_buf =
+    void* a_gk_gm0_gm10_gm11_grid_desc_dev_buf = workspace_buf.GetDeviceBuffer();
+    void* b_gk_gn0_gn10_gn11_grid_desc_dev_buf =
         static_cast<void*>(static_cast<unsigned char*>(workspace_buf.GetDeviceBuffer()) + 1024);
-    void* c_m0_m10_m11_n0_n10_n11_grid_desc_dev_buf =
+    void* c_gm10_bm0_bm1_gn10_bn0_bn1_grid_desc_dev_buf =
         static_cast<void*>(static_cast<unsigned char*>(workspace_buf.GetDeviceBuffer()) + 2048);
-    void* c_blockid_to_m0_n0_block_cluster_adaptor_dev_buf =
+    void* c_blockid_to_gm10_gn10_block_cluster_adaptor_dev_buf =
         static_cast<void*>(static_cast<unsigned char*>(workspace_buf.GetDeviceBuffer()) + 3072);
 
     const std::vector<size_t> vld  = {static_cast<size_t>(tunable->BlockSize), 1, 1};
     const std::vector<size_t> vgd1 = {static_cast<size_t>(tunable->BlockSize), 1, 1};
     const std::vector<size_t> vgd2 = {static_cast<size_t>(grid_size * tunable->BlockSize), 1, 1};
 
-    std::string program_name = "dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw.cpp";
+    std::string program_name = "dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw.cpp";
     std::string algo_name    = "implicit_gemm_conv_fwd_v4r4_nchw";
 
     std::string param = " -std=c++17 ";
     std::string network_config;
 
-    param += get_definition_string_from_types<TInWei, TAcc, TOut>() + " " +
-             get_definition_string_from_tunable(tunable) +
+    param += get_definition_string_from_types<TInWei, TAcc, TOut>() +
              " -DCK_PARAM_HAS_MAIN_KBLOCK_LOOP=" + std::to_string(hasMainKBlockLoop) +
-             " -DCK_PARAM_HAS_DOUBLE_TAIL_KBLOCK_LOOP=" + std::to_string(hasDoubleTailKBlockLoop);
-    network_config = get_network_config_string_from_types<TInWei, TAcc, TOut>() + "_" +
-                     get_network_config_string_from_tunable(tunable) + "_" +
-                     std::to_string(hasMainKBlockLoop) + "_" +
-                     std::to_string(hasDoubleTailKBlockLoop);
+             " -DCK_PARAM_HAS_DOUBLE_TAIL_KBLOCK_LOOP=" + std::to_string(hasDoubleTailKBlockLoop) +
+             " -DCK_PARAM_N0=" + std::to_string(N0) + " " +
+             get_definition_string_from_tunable(tunable);
+    network_config = get_network_config_string_from_types<TInWei, TAcc, TOut>() + "_V" +
+                     std::to_string(hasDoubleTailKBlockLoop) + "_" + std::to_string(N0) + "_" +
+                     get_network_config_string_from_tunable(tunable);
 
     std::vector<float> kernel1_times;
     std::vector<float> kernel2_times;
@@ -313,7 +334,7 @@ void device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc(
         KernelTimer timer1, timer2;
         std::string kernel_name;
 
-        kernel_name = "dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_prepare";
+        kernel_name = "dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw_prepare";
         auto network_config_1 = network_config + "_1";
 
         timer1.Start();
@@ -333,13 +354,13 @@ void device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc(
             in_left_pads[I1],
             in_right_pads[I0],
             in_right_pads[I1],
-            a_k_m0_m1_grid_desc_dev_buf,
-            b_k_n0_n1_grid_desc_dev_buf,
-            c_m0_m10_m11_n0_n10_n11_grid_desc_dev_buf,
-            c_blockid_to_m0_n0_block_cluster_adaptor_dev_buf);
-        timer1.End();
+            a_gk_gm0_gm10_gm11_grid_desc_dev_buf,
+            b_gk_gn0_gn10_gn11_grid_desc_dev_buf,
+            c_gm10_bm0_bm1_gn10_bn0_bn1_grid_desc_dev_buf,
+            c_blockid_to_gm10_gn10_block_cluster_adaptor_dev_buf);
+        timer2.End();
 
-        kernel_name           = "dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw";
+        kernel_name           = "dynamic_convolution_forward_implicit_gemm_v4r5_nchw_kcyx_nkhw";
         auto network_config_2 = network_config + "_2";
 
         timer2.Start();
@@ -347,10 +368,10 @@ void device_dynamic_convolution_forward_implicit_gemm_v4r4_nchw_kcyx_nkhw_olc(
             reinterpret_cast<const TInWei*>(wei_k_c_y_x_dev_buf.GetDeviceBuffer()),
             reinterpret_cast<const TInWei*>(in_n_c_hi_wi_dev_buf.GetDeviceBuffer()),
             reinterpret_cast<TOut*>(out_n_k_ho_wo_dev_buf.GetDeviceBuffer()),
-            (const void*)(a_k_m0_m1_grid_desc_dev_buf),
-            (const void*)(b_k_n0_n1_grid_desc_dev_buf),
-            (const void*)(c_m0_m10_m11_n0_n10_n11_grid_desc_dev_buf),
-            (const void*)(c_blockid_to_m0_n0_block_cluster_adaptor_dev_buf));
+            (const void*)(a_gk_gm0_gm10_gm11_grid_desc_dev_buf),
+            (const void*)(b_gk_gn0_gn10_gn11_grid_desc_dev_buf),
+            (const void*)(c_gm10_bm0_bm1_gn10_bn0_bn1_grid_desc_dev_buf),
+            (const void*)(c_blockid_to_gm10_gn10_block_cluster_adaptor_dev_buf));
         timer2.End();
 
         kernel1_times.push_back(timer1.GetElapsedTime());
